smokebot: correction to giturl emaail output

+++ Smokebot/smokebot.sh
@@ -1427,14 +1427,12 @@ email_build_status()
   fi
   if [ "$UPLOADRESULTS" == "1" ]; then
     echo "               status: https://pages.nist.gov/fds-smv/smokebot_status.html" >> $TIME_LOG
-    if [[ ! -e $WARNING_LOG && ! -e $ERROR_LOG ]]; then
-      if [ `whoami` == "smokebot" ];  then
-        echo  "***output guides and figures to Github"             > output/stage_GHupload
-        echo  ""                                                  >> output/stage_GHupload
-        $UploadGuidesGH                                          &>> output/stage_GHupload
-        GITURL=https://github.com/$GH_OWNER/$GH_REPO/releases/tag/$GH_SMOKEVIEW_TAG
-        echo "               guides: $GITURL"  >> $TIME_LOG
-      fi
+    if [[ `whoami` == "smokebot" ]] && [[ ! -e $WARNING_LOG ]] && [[ ! -e $ERROR_LOG ]]; then
+      echo  "***output guides and figures to Github"             > output/stage_GHupload
+      echo  ""                                                  >> output/stage_GHupload
+      $UploadGuidesGH                                          &>> output/stage_GHupload
+      GITURL=https://github.com/$GH_OWNER/$GH_REPO/releases/tag/$GH_SMOKEVIEW_TAG
+      echo "               guides: $GITURL"  >> $TIME_LOG
     fi
   fi
   echo "-------------------------------"   >> $TIME_LOG
